+++ HISTORY.md
@@ -25,8 +25,7 @@
     * new "get graphic effect" reporter
     * new "get pen attribute" reporter
     * new "write" command in pen category (used to be "label" in tools)
-    * added "neg" selector to monadic function reporter in "Operators" category
-    * added "log2" selector to monadic function reporter in "Operators" category
+    * added "-" (minus), "lg" (log2) and "2^" selectors to monadic function reporter in Operators
     * added "^" reporter (power of) in the Operators category
     * added "width" and "height" as attribute selectors of the OF primitive for the stage
     * special context-aware drop-downs for custom blocks
@@ -66,6 +65,9 @@
     * Catalan, thanks, Joan!
     * German
     * French
+
+### 2019-04-11
+* Blocks, Threads: Renamed monadic selectors: "neg" to "-" and "log2" to "lg", added "2^"
 
 ### 2019-04-10
 * Objects: took out MAP and FOREACH primitives (available in dev mode)